+++ tests/test_max_size_and_optional_assets.py
@@ -27,11 +27,10 @@ class TestMaxSizeAggregationCount(TestCase):
         source will be placed due to the minimization of the cost (which cost is this?, I assume
         operational cost) and installation cost. The placement behaviour is further tested in a
         second case by adding an optional ates and buffer. However, these 2 additional optional
-        assets should not be placed by the optmizer because of heat losses.
+        assets should not be placed by the optImizer because of heat losses.
 
         Checks:
-        - Check that heat_loss_efficiency of buffer is read from esdl and used to calculate
-            the heat loss of buffer
+        - Check the heat loss of buffer calculation
         - Check that source 1 is utilized and also placed
         - Check that source 2 is utilized and placed
         - Check that the geothermal source is not placed
@@ -62,8 +61,7 @@ class TestMaxSizeAggregationCount(TestCase):
         results = solution.extract_results()
         parameters = solution.parameters(0)
 
-        # Test that heat_loss_efficiency can be read from esdl
-        # and incorporated into heat loss calculation of buffer
+        # Test that if heat_loss_efficiency is incorporated into heat loss calculation of buffer
         esdl_asset = solution.esdl_assets[solution.esdl_asset_name_to_id_map["HeatStorage_74c1"]]
         np.testing.assert_allclose(
             parameters["HeatStorage_74c1.heat_loss_efficiency"],